pinch zooming now scales the RenderCanvas ImageView itself, rather than redrawing a zoomed image. This is quicker.
we renamed the Gestures class to GestureOverlay and made it a subclass of ImageView. An instance of this class is placed in the activity_main layout. we now listen for gestures on this class rather than the RenderCanvas because the process of scaling will bugger up gesture measurements

TODO: not quite right when moving from pinch zooming to single touch. not sure why this is. also chaining of pinch-zooming/scrolling still isn't quite right.