RenderCanvas.startRender() no longer called directly from RenderCanvas.animatedZoom(). instead, GestureOverlay.onDoubleTap() sets altered_canvas to true, like the other gesture methods do.
animatedZoom() calls GestureOverlay().block() and unblock() to cause the ACTION_UP handler to wait for the unblock before called startRender(). the ACTION_UP handler uses an AsyncTask() instance to do this

i've also added a time delay to this AsyncTask with the intention of making the UI more intuitive by not starting the render immediately after a canvas change. for example, after a double tap zoom, the user can no move the canvas before the offscreen image is destroyed and the render starting again.

the timings for this (defined by ON_UP_DELAY) may need some refinement